+++ index.js
@@ -87,34 +87,17 @@ async function factory (pkgName) {
      */
     exportData = async (params) => {
       const { get } = this.app.lib._
-      const { fs } = this.app.lib
       const { exportTo } = this.app.doboExtra
       const model = get(params, 'payload.data.name')
       const fields = get(params, 'payload.data.opts.fields')
-      const { id, file } = get(params, 'payload.data.download', {})
+      const file = get(params, 'payload.data.file', {})
       const options = {
         filter: get(params, 'payload.data.filter', {}),
         exportOpts: get(params, 'payload.data.exportOpts', []),
         opts: get(params, 'payload.data.opts', {}),
         fields
       }
-      options.filter.sort = 'id:1' // TODO: make this configurable
-      if (!this.app.sumba || !id) {
-        const result = await exportTo(model, file, options)
-        return result.file
-      }
-      const dmodel = this.app.dobo.getModel('SumbaDownload')
-      let dest
-      try {
-        await dmodel.updateRecord(id, { status: 'PROCESSING' })
-        const result = await exportTo(model, file, options)
-        dest = result.file
-        const { size } = fs.statSync(dest)
-        await dmodel.updateRecord(id, { size, status: 'COMPLETE' })
-      } catch (err) {
-        await dmodel.updateRecord(id, { status: 'FAIL' })
-      }
-      return dest
+      return await exportTo(model, file, options)
     }
 
     /**